fix(scripts): rename ingredient-data to .mjs; revert package.json type:module and CJS configs
