+++ e2e/localstorage.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test';
 import { mockProductApi } from '../tests/helpers/mock-api';
-import vermeiden from '../tests/fixtures/products/vermeiden.json' with { type: 'json' };
+import vermeiden from '../tests/fixtures/products/vermeiden.json';
 import gut from '../tests/fixtures/products/gut.json' with { type: 'json' };
 
 const SKIPPED_KEY = 'hashimoto-pcos-onboarding-skipped';
