+++ package.json
@@ -2,7 +2,6 @@
   "name": "hashimoto-pcos",
   "version": "0.1.0",
   "private": true,
-  "type": "module",
   "scripts": {
     "dev": "next dev",
     "dev:clean": "rm -rf .next && next dev",
