+++ postcss.config.js
@@ -1,4 +1,4 @@
-export default {
+module.exports = {
   plugins: {
     '@tailwindcss/postcss': {},
   },
